edit

+++ MFTEntry.h
@@ -2,12 +2,12 @@
 #define __MFT_ENTRY_H__
 #include "helper.h"
 using namespace std;
-int ReadMFTEntry(LPCWSTR  drive, uint32_t readPoint, BYTE entry[1024])
+int ReadMFTEntry(LPCWSTR  drive, uint64_t readPoint, BYTE entry[1024])
 {
     DWORD bytesRead;
     HANDLE device = NULL;
 
-    device = CreateFileW(drive,    // Drive to open
+    device = CreateFileW(drive, // Drive to open
         GENERIC_READ,           // Access mode
         FILE_SHARE_READ | FILE_SHARE_WRITE,        // Share Mode
         NULL,                   // Security Descrptori
@@ -31,25 +31,6 @@ int ReadMFTEntry(LPCWSTR  drive, uint32_t readPoint, BYTE entry[1024])
     {
     
         return 0;
-        // cout << "OFFSET:   0  1  2  3  4  5  6  7  8  9  A  B  C  D  E  F" << endl; 
-       
-        // for (int i = 0; i < 512; i++) {
-        //     if (i % 16 == 0) {
-        //         cout << endl;
-        //         cout << toHex(i/16) <<"0   :  ";
-        //     }
-             
-        //     cout << toHex(entry[i]) << " ";
-        // }
-        // cout << endl;
-        // for (int i = 512; i < 1024; i++){
-        //     if (i % 16 == 0) {
-        //         cout << endl;
-        //         cout << toHex(i/16) <<"0   :  ";
-        //     }
-             
-        //     cout << toHex(entry[i]) << " ";
-        // }
         
     }
     return 0;
